Make it possible to extend the size of the heap
Add extend call that will increase the size of the heap by the given
size. This is unsafe call and the caller must ensure that the new area
is valid.

That allows heap to be more memory efficient, consuming more memory only
when necessary, eg. os allocator might map additional virtual memory when
the allocation fails.